Move a lot of stuff to ThemeGenerator
